% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_linked_status_col}
\alias{add_linked_status_col}
\title{Add a column defining linked status}
\usage{
add_linked_status_col(.data)
}
\arguments{
\item{.data}{a data frame (or tibble) with a studyid column}
}
\value{
the same data frame plus a \code{linked_status} column
}
\description{
This function is really meant to be used in a pipeline as a quick shortcut to add the
linked status column
}
