% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-ages.R
\name{age_at_date}
\alias{age_at_date}
\title{Calculate the age of an invidividual on a certain date.}
\usage{
age_at_date(age_at_date, dobyyyy, dobmm)
}
\arguments{
\item{age_at_date}{date from which you want to calculate the age}

\item{dobyyyy}{year of birth.}

\item{dobmm}{month of birth.}
}
\description{
This function is setup to primarily to deal with the demographics file.
}
