% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{branch}
\alias{branch}
\title{Detection of branch points in a skeletonized image}
\usage{
branch(x)
}
\arguments{
\item{x}{skeletonized image}
}
\description{
Detection of branch points in a skeletonized image
}
\examples{
branch()
}

