% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-to-arrow-formats.R
\name{dat_to_arrow_formats}
\alias{dat_to_arrow_formats}
\alias{dat_to_parquet}
\alias{dat_to_arrow}
\alias{dat_to_feather}
\title{Convert .dat.gz files into arrow formats}
\usage{
dat_to_arrow_formats(
  data_path,
  data_dict,
  output_dir,
  arrow_format,
  col_types = NULL,
  col_select = NULL,
  overwrite = TRUE,
  data_format = c("fwf", "csv", "tsv", "csv2"),
  tz = "UTC",
  date_format = "\%AD",
  time_format = "\%AT",
  ...
)

dat_to_parquet(...)

dat_to_arrow(...)

dat_to_feather(...)
}
\arguments{
\item{data_path}{A path or a vector of paths to a \code{.dat.gz} file. If supplying a vector of paths,
they must share a common data dictionary.}

\item{data_dict}{A data.frame with \code{start}, \code{stop} and \code{name} columns}

\item{output_dir}{path to where you want to save the output file}

\item{arrow_format}{must be one of parquet, arrow or feather}

\item{col_types}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be imputed from \code{guess_max} rows
on the input interspersed throughout the file. This is convenient (and
fast), but not robust. If the imputation fails, you'll need to increase
the \code{guess_max} or supply the correct types yourself.

Column specifications created by \code{\link[=list]{list()}} or \code{\link[readr:cols]{cols()}} must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[readr:cols]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
\itemize{
\item c = character
\item i = integer
\item n = number
\item d = double
\item l = logical
\item f = factor
\item D = date
\item T = date time
\item t = time
\item ? = guess
\item _ or - = skip

By default, reading a file without a column specification will print a
message showing what \code{readr} guessed they were. To remove this message,
set \code{show_col_types = FALSE} or set `options(readr.show_col_types = FALSE).
}}

\item{col_select}{A vector of column names}

\item{overwrite}{logical; should existing destination files be overwritten?}

\item{data_format}{the format of the input data. Default is \code{"fwf"}, other choices
are \code{"csv"}, \code{"csv2"}, \code{"tsv"}}

\item{tz}{what timezone should datetime fields use? Default UTC. This is recommended
to avoid timezone pain, but remember that the data is in UTC when doing analysis. See
OlsonNames() for list of available timezones.}

\item{date_format}{date format for columns where date format is not specified in \code{col_types}}

\item{time_format}{time format for columns where time format is not specified in \code{col_types}}

\item{...}{passed to one or arrows writing functions depending on the value of \code{arrow_format}}
}
\description{
Function to convert into formats provided by Apache Arrow.
These functions will overwrite files of the same name by default.
Convenience wrappers are provided for each arrow format.
}
\section{Functions}{
\itemize{
\item \code{dat_to_parquet}: Convert to parquet

\item \code{dat_to_arrow}: Convert to arrow

\item \code{dat_to_feather}: Convert to feather
}}

