% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dat.R
\name{delete_rtmp_dirs}
\alias{delete_rtmp_dirs}
\title{Delete all content from any temp directory}
\usage{
delete_rtmp_dirs(are_you_sure = FALSE)
}
\arguments{
\item{are_you_sure}{logical Defaults to FALSE so that a user does not do this by accident.}
}
\description{
This function helps clean up after vroom and other functions that use the temp directory
to store transient files. Sometimes this can balloon in size and it is useful run this command to
clean things up.
}
