% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{dipr_create_targets_project}
\alias{dipr_create_targets_project}
\title{An opinionated template for package-like targets project}
\usage{
dipr_create_targets_project(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[usethis:use_description]{usethis::use_description}}
  \describe{
    \item{\code{fields}}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}
    \item{\code{check_name}}{Whether to check if the name is valid for CRAN and throw an
error if not}
    \item{\code{roxygen}}{If \code{TRUE}, sets \code{RoxygenNote} to current roxygen2 version}
  }}
}
\description{
This function will create a DESCRIPTION file, a \verb{_targets.R} template and an \code{R}
directory to store function files.
}
