% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{dipr_document_output_groups}
\alias{dipr_document_output_groups}
\title{Convenience function to generate templates for documenting export datasets}
\usage{
dipr_document_output_groups(path, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{...}{
  Arguments passed on to \code{\link[base:list.files]{base::list.files}}
  \describe{
    \item{\code{pattern}}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
    \item{\code{all.files}}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned (following Unix-style visibility,
    that is files whose name does not start with a dot).  If \code{TRUE},
    all file names will be returned.}
    \item{\code{full.names}}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}
    \item{\code{recursive}}{logical.  Should the listing recurse into directories?}
    \item{\code{ignore.case}}{logical.  Should pattern-matching be case-insensitive?}
    \item{\code{include.dirs}}{logical.  Should subdirectory names be included in
    recursive listings?  (They always are in non-recursive ones).}
    \item{\code{no..}}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}
  }}
}
\description{
The usage of this function is to print templated documentation to the console
and then copy and paste into the export document create with \code{dipr_use_export_doc}.
This function works only on csv exports.
}
