% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dipr_examples}
\alias{dipr_examples}
\alias{dipr_example}
\title{Get path to dipr example files}
\usage{
dipr_examples()

dipr_example(path)
}
\arguments{
\item{path}{A path to the data file}
}
\description{
\code{dipr} provides an example of compressed
fixed width file and a data dictionary. This data
is from the \code{dplyr} package and contains no real personal
information. These functions are modelled heavily of similar
functions in the vroom package.
}
\examples{

## Get all available data files
dipr_examples()

## Use that to find the path to a file
dipr_example("starwars-dict.txt")
}
