% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{dipr_icd_categories}
\alias{dipr_icd_categories}
\alias{dipr_icd9_categories}
\alias{dipr_icd10_categories}
\title{Assign a category to a vector of icd codes}
\usage{
dipr_icd_categories()

dipr_icd9_categories()

dipr_icd10_categories()
}
\description{
The dipr_icd_categories functions have been deprecated. Development on the hawkeye package
has superceded this functionality. In particular the *_lookup family of functions provide a
wider range of categories and icd codes.
}
\author{
Craig Hutton (\code{craig.hutton@gov.bc.ca})
}
