% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipr-update.R
\name{dipr_update}
\alias{dipr_update}
\title{Update the dipr package from gitlab}
\usage{
dipr_update(ref = NULL, ...)
}
\arguments{
\item{ref}{A tag of specific commit hash. Default to NULL which is the most recent release. "dev" will
install at the most recent commit.}

\item{...}{passed to install_sre_gitlab}
}
\description{
This function is a thin wrapper around \code{install_sre_gitlab} and is specific to
updating dipr. The default behaviour of this function is to only install if your
current version needs updating.
}
