% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{dipr_use_export_doc}
\alias{dipr_use_export_doc}
\title{Use the template export documentation}
\usage{
dipr_use_export_doc(template = "export-doc.md", ...)
}
\arguments{
\item{template}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}

\item{...}{further arguments passed to usethis::use_template}
}
\description{
This will open and save a markdown file with pre-populated
headings and sections to aid in generating output documentation.
This is not official documentation but rather a useful guide to
meet the standard for the output checkers. This function will
automatically output to "out/export" and number sequentially
the export folders.
}
\examples{
\dontrun{
dipr_use_export_doc()
}
}
