% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-to-arrow-formats.R
\name{dipr_write_parquet}
\alias{dipr_write_parquet}
\title{Write a parquet file in the SRE}
\usage{
dipr_write_parquet(x, sink, overwrite = TRUE, ...)
}
\arguments{
\item{x}{An \link[=Table]{arrow::Table}, or an object convertible to it.}

\item{sink}{an \link[=OutputStream]{arrow::io::OutputStream} or a string which is interpreted as a file path}

\item{overwrite}{logical; should existing destination files be overwritten?}

\item{...}{passed to \code{arrow::write_parquet}}
}
\description{
Wraps \code{arrow::write_parquet} but first writes to disk then moves to remote drive.
}
