% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmat.R
\name{distmat}
\alias{distmat}
\title{Calculate distance matrics}
\usage{
distmat(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x coordinate of the first point}

\item{y1}{y coordinate of the first point}

\item{x2}{x coordinate of the second point}

\item{y2}{y coordinate of the second point}
}
\description{
Calculate distance matrics
}
\examples{
distmat()
}

