% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ending.R
\name{ending}
\alias{ending}
\title{Detection of end points in a skeletonized image}
\usage{
ending(x)
}
\arguments{
\item{x}{skeletonized image}
}
\description{
Detection of end points in a skeletonized image
}
\examples{
ending()
}

