% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dict.R
\name{extract_path_name}
\alias{extract_path_name}
\title{A function to extract the data name from a path}
\usage{
extract_path_name(path)
}
\description{
This function depends on the path being a .dat.gz file
}
