% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-arrow.R
\name{filter_across}
\alias{filter_across}
\title{Search across columns in a data.frame or an arrow .Dataset for rows matching a given pattern.}
\usage{
filter_across(.data, code, cols, partial = FALSE)
}
\arguments{
\item{.data}{An object of either class data.frame or arrow .Dataset}

\item{code}{A character string representing a search term.}

\item{cols}{A vector of column names to filter on}

\item{partial}{A logical determining when partial matching should occur}
}
\description{
Searches across columns in a data object for rows matching a given pattern.
}
\details{
\itemize{
\item Function currently returns a "data.frame" if passed a "data.frame".  If passed a .Dataset or arrow_dplyr_query,
the result will be an object of class arrow_dplyr_query which can preceed arrow::collect().
}
}
