% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{filter_linked}
\alias{filter_linked}
\title{Wrapper to filter out unlinked studyids from either a data.frame or a Dataset}
\usage{
filter_linked(.data, ...)
}
\arguments{
\item{.data}{a data frame (or tibble) or arrow Dataset with a studyid column}

\item{...}{
  Arguments passed on to \code{\link[=add_linked_status_col]{add_linked_status_col}}
  \describe{
    \item{\code{studyid_col}}{the name of the study id column. Defaults to "studyid"}
  }}
}
\description{
This is a convenience function which automatically filter for only "linked" studyids
as defined by \code{add_linked_status()}.
}
\examples{

filter_linked(data.frame(studyid = c("sxxxx", "uxxxx")))
}
