% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-arrow.R
\name{flag_across}
\alias{flag_across}
\title{Add a logical flag (new column) based on a pattern search across columns}
\usage{
flag_across(.data, code = NULL, cols = NULL, partial = FALSE, flag_name = NULL)
}
\arguments{
\item{.data}{An object of either class data.frame or arrow .Dataset}

\item{code}{A character string representing a search term.}

\item{cols}{a vector of column names to search in}

\item{partial}{A logical determining when partial matching should occur}

\item{flag_name}{the name of the new column to create}
}
\value{
An object the same class as the input (methods for \code{data.frame}, \code{Dataset}, \code{arrow_dplyr_query}),
with a new logical column
}
\description{
Add a logical flag (new column) based on a pattern search across columns
}
\seealso{
\code{\link[=filter_across]{filter_across()}}
}
