% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmf2tif.R
\name{fmf2tif}
\alias{fmf2tif}
\title{Convert fmf file format to tif format}
\usage{
fmf2tif(filename, skip = 0)
}
\arguments{
\item{filename}{fmf file name}

\item{skip}{Skip frames}
}
\description{
Convert fmf file format to tif format
}
\examples{
fmf2tif()

}

