% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCC.R, R/SAD.R, R/SimpleCrossCorrelation.R
\name{FNCC}
\alias{FNCC}
\alias{NCC}
\alias{SAD}
\title{Template matching}
\usage{
FNCC(trg, tmp)

SAD(trg, tmp)

NCC(trg, tmp)
}
\arguments{
\item{trg}{A target image of Image object or an array.}

\item{tmp}{A template image of Image object or an array.}
}
\value{
Returns a grayscale image.
}
\description{
These functions perform template matching using SAD (Sum of absolute differences), NCC (Normalized cross correlation), or FNCC (fast normalized cross correlation).
}
\examples{
FNCC()
SAD()
NCC()
}

