% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-table-paths.R
\name{get_core_dict_path}
\alias{get_core_dict_path}
\title{Get paths of data dictionaries}
\usage{
get_core_dict_path(
  core_version,
  core_dict_name,
  base_path = "R:/DATA/core-snapshot"
)
}
\arguments{
\item{core_version}{Core version of tables typically formatted as a YYYYMMDD string}

\item{core_dict_name}{a string of the dataset name.}

\item{base_path}{Defaults to "R:/DATA/core-snapshot"}
}
\description{
Get paths of data dictionaries
}
