% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipr-update.R
\name{get_gitlab_sre_repos}
\alias{get_gitlab_sre_repos}
\title{Return a data.frame of all repos available to the project you are currently in}
\usage{
get_gitlab_sre_repos(token = set_gitlab_credentials()$password)
}
\arguments{
\item{token}{Credentials from gitlab, accessed via: \code{set_gitlab_credentials()$password}}
}
\description{
Access information on the repo in the current project in a data.frame.
}
\examples{
\dontrun{
get_gitlab_sre_repos(set_gitlab_credentials()$password)
}
}
