% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Histogram of pixel intensitie}
\usage{
histogram(x, breaks = 512, xlim = c(0, 1), ylim = NULL, main = "",
  col = "black")
}
\arguments{
\item{x}{An image of Image or imagedata class or an array.}

\item{breaks}{A single number giving the number of cells for the histogram.}

\item{xlim, ylim}{the range of x and y values with sensible defaults.}

\item{col}{A colour to be used to fill the bars.}
}
\description{
This function displays a histogram of pixel intensities from images. This is a wrapper function for hist.
}
\examples{
histogram()
}

