% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-addin.R
\name{insert_bcgov_cc_header}
\alias{insert_bcgov_cc_header}
\title{Add the boilerplate Creative Commons Attribution 4.0 (CC-BY) header to the
top of a source code file}
\usage{
insert_bcgov_cc_header(file, year = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{file}{Path to the file}

\item{year}{The year the licence should apply (Default current year)}
}
\description{
Add the boilerplate Creative Commons Attribution 4.0 (CC-BY) header to the
top of a source code file
}
\author{
Andy Teucher
}
