% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipr-update.R
\name{install_sre_gitlab}
\alias{install_sre_gitlab}
\title{Install a package from the SRE gitlab.}
\usage{
install_sre_gitlab(repo, ref = NULL, quiet = TRUE)
}
\arguments{
\item{repo}{The name of the repo.}

\item{ref}{A tag of specific commit hash. Default to NULL which is the most recent release. "dev" will
install at the most recent commit.}

\item{quiet}{If \code{TRUE}, suppress output.}
}
\description{
This function will install packages other than dipr from the SRE gitlab. The default behaviour
is to install releases.
}
\examples{
\dontrun{
install_sre_gitlab("hawkeye")
}
}
