% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocwa_branch_export}
\alias{ocwa_branch_export}
\title{Prepare repo for import into the SRE through OCWA}
\usage{
ocwa_branch_export(branch = "ocwa-import", ask = TRUE)
}
\arguments{
\item{branch}{branch name to prepare for OCWA. Default \code{"ocwa-import"}}

\item{ask}{should the function stop and ask for verification to proceed? Default \code{TRUE}}
}
\value{
hash of commit
}
\description{
Removes files listed in \verb{_ocwaignore} and processes \code{README.md}
so that it is not using images or URLs that can't be accessed in the SRE.
}
\details{
This function:
\enumerate{
\item Creates a new branch
\item Removes files that can't be imported - these are listed in the \verb{_ocwaignore} file in the root of the repo
\item Cleans \code{README.md} to comment out references to images and links that won't be available in the SRE.
\item Commits the changes from 2 and 3 to the new branch and pushing that to GitHub. This branch can then be used as the basis for an import into OCWA.
}
}
