% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoColor.R
\name{pseudoColor}
\alias{pseudoColor}
\title{Moment calculation}
\usage{
pseudoColor(x, min = 0, max = 255)
}
\arguments{
\item{x}{An Image object or a matrix.}

\item{min}{Minimum color range.}

\item{max}{Maximum color range.}
}
\description{
Moment calculation
}
\examples{
pseudoColor()
}

