% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2int.R
\name{raw2int}
\alias{raw2int}
\title{A raw2int Function}
\usage{
raw2int(rawdata)
}
\arguments{
\item{rawdata}{raw value}
}
\description{
This function converts raw to integer.
}
\examples{
raw2int()
}
\keyword{raw}

