% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2intR.R, R/raw2intcpp.R
\name{raw2intR}
\alias{raw2intR}
\title{2 byte raw to integer conversion}
\usage{
raw2intR(obj)

raw2intR(obj)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}

\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
2 byte raw to integer conversion

2 byte raw to integer conversion
}
\examples{
raw2intcpp()
raw2intcpp()
}

