% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2num.R
\name{raw2num}
\alias{raw2num}
\title{raw to integer conversion}
\usage{
raw2num(rawdata)
}
\arguments{
\item{rawdata}{A target image of Image object or an array.}
}
\description{
raw to integer conversion
}
\examples{
raw2num()
}

