% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dict.R
\name{read_health_dict}
\alias{read_health_dict}
\title{Read Health Data Dictionaries}
\usage{
read_health_dict(path, sheet, ...)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{...}{arguments passed to \code{readxl::read_excel}}
}
\description{
Read into R pre-formatted popdata data dictionaries. These files are in .xlsx format and therefore
require both a path and sheet argument. The rest of the function is a thin wrapper around reaxl::read_excel
with some formatting taking place.
}
