% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAVI.R
\name{readAVI}
\alias{readAVI}
\title{A readAVI Function}
\usage{
readAVI(filepath, start = 1, end = 0, skip = 0, getFrames = F,
  crop = c(0, 0, 0, 0), silent = F)
}
\arguments{
\item{filepath}{A caracter string of the path to the file. Required.}

\item{start}{An integer of the start frame. Default = 1.}

\item{end}{An integer of the end frame. Default = 0 (last frame of the file).}

\item{skip}{An integer of skip frame. Default = 0 (read every frame).}

\item{getFrames}{Return number of frames. Default = False.}

\item{crop}{An integer vector depicting the xy coordinate of the top-left corner and the bottom-right corner you want to crop.}

\item{silent}{Whether or not show message. Default = False.}
}
\description{
This function allows you to load a grayscale AVI file.
}
\keyword{AVI}

