% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFMF.R
\name{readFMF}
\alias{readFMF}
\title{A readFMF Function}
\usage{
readFMF(filepath, start = 1, end = 0, skip = 0, crop = c(0, 0, 0, 0),
  frames = NULL, getFrames = F)
}
\arguments{
\item{filepath}{A caracter string of the path to the file. Required.}

\item{start}{An integer of the start frame. Default = 1.}

\item{end}{An integer of the end frame. Default = 0 (last frame of the file).}

\item{skip}{An integer of skip frame. Default = 0 (read every frame).}

\item{crop}{An integer vector depicting the xy coordinate of the top-left corner and the bottom-right corner you want to crop.}

\item{frames}{A integer vector of frames to be read. Defaul = NULL.}

\item{getFrames}{Return number of frames. Default = False.}
}
\description{
This function allows you to load a grayscale FMF file.
}
\examples{
readFMF()
}
\keyword{FMF}

