% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLsm.R
\name{readLsm}
\alias{readLsm}
\title{A readLsm Function}
\usage{
readLsm(filename)
}
\arguments{
\item{filepath}{A caracter string of the path to the file. Required.}
}
\description{
This function allows you to load a grayscale AVI file.
}
\examples{
readLsm()
}
\keyword{lsm}

