% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTIFF2.R
\name{readTIFF2}
\alias{readTIFF2}
\title{A readTIFF2 Function}
\usage{
readTIFF2(filename, start = 1, end = 0, crop = c(0, 0, 0, 0),
  frames = NULL, getFrames = F, intensity = F)
}
\arguments{
\item{start}{An integer of the start frame. Default = 1.}

\item{end}{An integer of the end frame. Default = 0 (last frame of the file).}

\item{crop}{Not implemented yet.}

\item{frames}{A vector specifying which frames to be loaded.}

\item{getFrames}{Return number of frames. Default = False.}

\item{intensity}{Whether or not return mean intensity of each frame. Default = False.}

\item{filepath}{A caracter string of the path to the file. Required.}
}
\description{
This function allows you to load a TIFF file.
}
\examples{
readTIFF2()
}
\keyword{tiff}

