% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-prefs.R
\name{restore_rstudio_prefs}
\alias{restore_rstudio_prefs}
\title{Copy saved RStudio preferences from PopData Gitlab repo to you local SRE
machine}
\usage{
restore_rstudio_prefs(folder, repo = "using-sre-gitlab")
}
\arguments{
\item{folder}{The folder in which your preferences are stored in the repo.
For a minimal recommended setup, use \code{"rstudio-prefs/generic"} as a starting point.}

\item{repo}{The name of the GitLab repo in which you have stored your
preferences. Default \code{"using-sre-gitlab"}}
}
\value{
a character vector of file paths of the files that were copied to
your machine (invisibly).
}
\description{
DSP R Users are storing their Rstudio preferences in the repo at:
\url{https://projectsc.popdata.bc.ca/shares/using-sre-gitlab}, in a named user
folder under \code{rstudio-prefs} (See the
\href{https://projectsc.popdata.bc.ca/shares/using-sre-gitlab/-/tree/develop/rstudio-prefs\%generic}{generic}
folder for a minimal recommended starting point, or the
\href{https://projectsc.popdata.bc.ca/shares/using-sre-gitlab/-/tree/develop/rstudio-prefs\%2Fandy}{andy}
folder for a more comprehensive example). Your local RStudio preferences are
located in C:/Users/[your-user-name]/AppData/Roaming/Rstudio. When you have
an instance of Rstudio set up to your liking, copy these to your named folder
in the repo of your choice, then use this function to restore those
preferences to a new SRE machine.
}
\examples{
\dontrun{
 restore_rstudio_prefs(folder = "rstudio-prefs/andy")
}
}
