% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollmeanimg.R
\name{rollmeanimg}
\alias{rollmeanimg}
\title{Rolling mean calculation for images}
\usage{
rollmeanimg(obj, n)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{n}{Window size.}
}
\description{
Rolling mean calculation for images
}
\examples{
rollmeanimg()
}

