% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rollmeanimgC}
\alias{rollmeanimgC}
\title{Image roll mean}
\usage{
rollmeanimgC(obj, dim, n)
}
\arguments{
\item{obj}{An array.}

\item{dim}{Int vector.}

\item{n}{window.}
}
\description{
Image roll mean
}

