% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollmedianimg.R
\name{rollmedianimg}
\alias{rollmedianimg}
\title{Rolling median calculation for images}
\usage{
rollmedianimg(obj, n)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{n}{Window size.}
}
\description{
Rolling median calculation for images
}
\examples{
rollmedianimg()
}

