% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-utils.R
\name{set_gitlab_credentials}
\alias{set_gitlab_credentials}
\title{Set gitlab credentials from R}
\usage{
set_gitlab_credentials(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[credentials:http_credentials]{credentials::git_credential_ask}}
  \describe{
    \item{\code{save}}{in case the user is prompted for credentials, attempt to
remember them.}
    \item{\code{verbose}}{print errors from \verb{git credential} to stdout}
  }}
}
\description{
Running this function raises the Windows Git Credential Manager to enter
the GitLat PAT. Accessing the GitLab PAT requires following these steps:
\itemize{
\item Go to this website: \url{https://projectsc.popdata.bc.ca/}.
\item Scroll down and click on "Sign in with Pop Data SSO".
\item Next go to this site: \url{https://projectsc.popdata.bc.ca/profile/personal_access_tokens}
\item Create a new Personal Access Token enabling each scope and typing in your username. This should create a new personal access token. Those are your "git credentials". Save those on your "U" drive along with your user name.
}
}
