% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfeatures.R
\name{sfeatures}
\alias{sfeatures}
\title{Moment calculation}
\usage{
sfeatures(obj, ref = 0)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Moment calculation
}
\examples{
sfeatures()
}

