% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfeaturesC}
\alias{sfeaturesC}
\title{Extract features}
\usage{
sfeaturesC(obj, ref, nx, ny, nz)
}
\arguments{
\item{obj}{An array.}

\item{ref}{A matrix.}

\item{nx}{x dimension.}

\item{ny}{y dimension.}

\item{nz}{z dimension.}
}
\description{
Extract features
}

