% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepC.R
\name{ssweep}
\alias{ssweep}
\title{Fast sweep function}
\usage{
ssweep(obj, ref, op)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}

\item{op}{An operator.}
}
\description{
Fast sweep function
}
\examples{
ssweep()
}

