% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackDistance.R
\name{trackDistance}
\alias{angularVelocity}
\alias{msd}
\alias{trackDistance}
\alias{trackSpeed}
\alias{tracking}
\title{Template matching}
\usage{
trackSpeed(dist, bin = 5, interval = 1)
}
\arguments{
\item{dist}{Traveled distance (e.g. the output of trackDistance).}

\item{bin}{Window size for averaging.}

\item{interval}{Time intervals.}

\item{mask}{Labeled binary movie.}

\item{maxdist}{Maximum distance for object association.}

\item{unit}{Unit length.}

\item{size}{Maximum size difference for objects association.}

\item{xy}{A matrix of object coordinates.}

\item{xy}{A target image of Image object or an array.}
}
\value{
Returns a grayscale image.
}
\description{
These functions are for tracking and analyzing objects in movies. tracking calculates various values related to object motion. trackDistance calculates distance traveled by objects during the specified time window. trackSpeed calculates average speed of objects during the specified time window. angularVelocity calculates angular velocity of moving objects. msd calculates mean square diviation of object trajectories.
}
\examples{
tracking()
}

