% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angularVelocity.R, R/msd.R, R/trackDistance.R, R/trackSpeed.R, R/tracking.R
\name{angularVelocity}
\alias{angularVelocity}
\alias{msd}
\alias{trackDistance}
\alias{trackSpeed}
\alias{tracking}
\title{Object tracking}
\usage{
angularVelocity(xy, bin = 5, interval = 1)

msd(xy)

trackDistance(xy, unit = 1)

trackSpeed(dist, bin = 5, interval = 1)

tracking(rdir, mask = NULL, w = 0, h = 0, maxdist = 20, bin = 3,
  interval = 0.1, unit = 1, size = NULL, ftrs = NULL)
}
\arguments{
\item{xy}{A matrix of object coordinates.}

\item{bin}{Window size for averaging.}

\item{interval}{Time intervals.}

\item{unit}{Unit length.}

\item{dist}{Traveled distance (e.g. the output of trackDistance).}

\item{mask}{Labeled binary movie.}

\item{maxdist}{Maximum distance for object association.}

\item{size}{Maximum size difference for objects association.}

\item{xy}{A target image of Image object or an array.}
}
\description{
These functions are for tracking and analyzing objects in movies. tracking calculates various values related to object motion. trackDistance calculates distance traveled by objects during the specified time window. trackSpeed calculates average speed of objects during the specified time window. angularVelocity calculates angular velocity of moving objects. msd calculates mean square diviation of object trajectories.
}
\examples{
angularVelocity()
msd()
trackDistance()
trackSpeed()
tracking()
}

