% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_linked_status_col}
\alias{add_linked_status_col}
\title{Add a column defining linked status}
\usage{
add_linked_status_col(.data, studyid_col = "studyid")
}
\arguments{
\item{.data}{a data frame (or tibble) or arrow Dataset with a studyid column}

\item{studyid_col}{the name of the study id column. Defaults to "studyid"}
}
\value{
the same data frame plus a \code{linked_status} column
}
\description{
This function is really meant to be used in a pipeline as a quick shortcut to add the
linked status column
}
