% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{dipr_icd_categories}
\alias{dipr_icd_categories}
\alias{dipr_icd9_categories}
\alias{dipr_icd10_categories}
\alias{clean_icd_codes}
\title{Assign a category to a vector of icd codes}
\usage{
dipr_icd_categories(
  icd_codes = NULL,
  icd_version = 10,
  include_code_ranges = FALSE,
  mortality = FALSE,
  print_defs_short = FALSE,
  print_defs_long = FALSE,
  export_defs = FALSE
)

dipr_icd9_categories(...)

dipr_icd10_categories(...)

clean_icd_codes(icd_codes)
}
\arguments{
\item{icd_codes}{vector of icd codes}

\item{icd_version}{version of the icd code system}

\item{include_code_ranges}{include codes used to assign category in returned vector.}

\item{mortality}{default is FALSE. Set to TRUE if working with mortality data}

\item{print_defs_short}{print a short definition}

\item{print_defs_long}{print a long definition}

\item{export_defs}{write out a csv. Defaults to FALSE.}

\item{...}{arguments passed to \code{dipr_icd_categories}}
}
\description{
Vectorized function that assigns categories to icd codes. Thin wrappers are provided for
convenience for ICD-9 and ICD-10 codes respectively. Other than the \code{icd_version} argument,
each wrapper accepts the same arguments as the core function.
}
\details{
\itemize{
\item ICD-10 categories were defined according to the official 2019 WHO ICD-10 website: https://icd.who.int/browse10/2019/en
\item ICD-9 categories were defined according to the BC Ministry of Health's ICD-9 Diagnostic Code Descriptions
website: https://www2.gov.bc.ca/gov/content/health/practitioner-professional-resources/msp/physicians/diagnostic-code-descriptions-icd-9
}
}
\section{Functions}{
\itemize{
\item \code{dipr_icd9_categories}: Thin wrapper around dipr_icd_categories for only ICD-9 codes

\item \code{dipr_icd10_categories}: Thin wrapper around dipr_icd_categories for only ICD-10 codes

\item \code{clean_icd_codes}: Method to clean ICD-9 and ICD-10 codes
}}

\examples{

## ICD-9
dipr_icd9_categories("C19")

## ICD-10
dipr_icd10_categories("U07.1")
}
\author{
Craig Hutton (\code{craig.hutton@gov.bc.ca})
}
