% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipr-update.R
\name{dipr_update}
\alias{dipr_update}
\title{Update the dipr package from gitlab}
\usage{
dipr_update(quiet = TRUE, ref = NULL)
}
\arguments{
\item{quiet}{If \code{TRUE}, suppress output.}

\item{ref}{A tag of specific commit hash}
}
\description{
This function clones the dipr gitlab repo and install the package. This
is a useful function update the package and saves the user from having to
manually accomplish each step.
}
