% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCC.R
\name{FNCC}
\alias{FNCC}
\title{An FNCC Function}
\usage{
FNCC(a, b)
}
\arguments{
\item{a}{A target image of Image object or an array.}

\item{b}{A template image of Image object or an array.}
}
\value{
Returns a grayscale image.
}
\description{
These functions perform template matching using SAD (Sum of absolute differences), NCC (Normalized cross correlation), or FNCC (fast normalized cross correlation).
}
\examples{
demo(TemplateMatchingDemo)
}
\keyword{Template}
\keyword{matching}

