% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-table-paths.R
\name{get_core_dat_path}
\alias{get_core_dat_path}
\title{Get the dat path of a particular dataset for a given core table name}
\usage{
get_core_dat_path(
  core_version,
  core_table_name,
  pattern = "\\\\.dat.gz$",
  base_path = "R:/DATA/core-snapshot"
)
}
\arguments{
\item{core_version}{Core version of tables typically formatted as a YYYYMMDD string}

\item{core_table_name}{a string of the data set name}

\item{pattern}{Defaults to dat.gz but can be extended for more extended using regex for more
flexibility}

\item{base_path}{Defaults to "R:/DATA/core-snapshot"}
}
\description{
Return the paths of all data within a version that match a table name
}
