% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ocwa_branch_export}
\alias{ocwa_branch_export}
\title{Prepare repo for import into the SRE through OCWA}
\usage{
ocwa_branch_export(branch = "ocwa-import")
}
\arguments{
\item{branch}{branch name to prepare for OCWA. Default \code{"ocwa-import"}}
}
\value{
hash of commit
}
\description{
Removes files listed in \verb{_ocwaignore} and processes \code{README.md}
so that it is not using images or URLs that can't be accessed in the SRE.
}
\details{
Then checks out a branch in \code{{dipr}} and commits those changes to that branch.
Then use that branch to import into OCWA.
}
