% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dat.R
\name{read_dat}
\alias{read_dat}
\title{Read compressed files in R}
\usage{
read_dat(data_path, data_dict, use_cache = FALSE, cache_type = "fst",
  cache_dir = NULL, as.data.table = TRUE, col_select = NULL,
  col_types = NULL, ...)
}
\arguments{
\item{data_path}{A path to a \code{.dat.gz} file}

\item{data_dict}{A data.frame with \code{start}, \code{stop} and \code{name} columns}

\item{use_cache}{A logical value if you want to use a cache. This create a \code{.dipr} directory
in your project and stores data there. Defaults to FALSE.}

\item{cache_type}{A string of the file type you want to store as your cache. Possible values
are "fst", "csv", or "parquet"}

\item{cache_dir}{Directory where you want you to store your cache. Defaults to \code{NULL} which results in
a \code{.dipr} directory in your project. This can also be set by the environment variable \code{DIPR_CACHE_PATH}}

\item{as.data.table}{A logical value if you want a data.table object returned.}

\item{col_select}{A vector of column names}

\item{col_types}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be imputed from the first 1000 rows
on the input. This is convenient (and fast), but not robust. If the
imputation fails, you'll need to supply the correct types yourself.

If a column specification created by \code{\link[readr:cols]{cols()}}, it must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[readr:cols_only]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
c = character, i = integer, n = number, d = double,
l = logical, f = factor, D = date, T = date time, t = time, ? = guess, or
\code{_}/\code{-} to skip the column.}

\item{...}{arguments passed to \code{vroom::vroom_fwf}}
}
\description{
Unzip and read into memory (using vroom). The first time you call this function, the data
is cached in a folder called \code{.dipr} in a format specified by the cache_type argument.
This function expects a data dictionary with the following columns:
\itemize{
\item start
\item stop
\item name
}
}
