% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dict.R
\name{read_nflt}
\alias{read_nflt}
\title{Read data dictionaries in nflt format
This function will be quite fragile to changes in the data format. It assumes four columns}
\usage{
read_nflt(path)
}
\description{
Read data dictionaries in nflt format
This function will be quite fragile to changes in the data format. It assumes four columns
}
