% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTIFF2.R
\name{readTIFF2}
\alias{readTIFF2}
\title{A readTIFF2 Function}
\usage{
readTIFF2(filename, start = 1, end = 0, crop = c(0, 0, 0, 0),
  frames = NULL, getFrames = F, intensity = F)
}
\arguments{
\item{start}{An integer of the start frame. Default = 1.}

\item{end}{An integer of the end frame. Default = 0 (last frame of the file).}

\item{crop}{An integer vector depicting the xy coordinate of the top-left corner and the bottom-right corner you want to crop.}

\item{getFrames}{Return number of frames. Default = False.}

\item{filepath}{A caracter string of the path to the file. Required.}

\item{silent}{Whether or not show message. Default = False.}
}
\description{
This function allows you to load a TIFF file.
}
\examples{
readTIFF2()
}
\keyword{tiff}

