% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sweepC}
\alias{sweepC}
\title{sweep array with matrix}
\usage{
sweepC(obj, m, dim, op)
}
\arguments{
\item{obj}{An array.}

\item{m}{A matrix.}

\item{dim}{Dimension.}

\item{op}{Operation, either - or *}
}
\description{
sweep array with matrix
}

